+++ jaxley/__init__.py
@@ -9,7 +9,6 @@ from jaxley.connect import (
     sparse_connect,
 )
 from jaxley.integrate import integrate
-from jaxley.io.morph_utils import morph_connect, morph_delete
 from jaxley.io.swc import read_swc
 from jaxley.modules import *
 from jaxley.optimize import ParamTransform
@@ -17,8 +16,6 @@ from jaxley.stimulus import datapoint_to_step_currents, step_current
 
 __all__ = [
     "read_swc",
-    "morph_connect",
-    "morph_delete",
     "Module",
     "Branch",
     "Cell",