+++ jaxley/io/swc.py
@@ -471,6 +471,11 @@ def read_swc(
         A `Cell` object."""
 
     if backend == "custom":
+        warn(
+            "You set `backend='custom'. This option will be removed in the future. "
+            "If you are experiencing issues with the default SWC reader, please open "
+            "a `New issue` on GitHub: https://github.com/jaxleyverse/jaxley/issues"
+        )
         return read_swc_custom(
             fname,
             ncomp=ncomp,